S2US2, 8 + 18 implemented
S2US2: As a player I do not want the camera to rotate when I rotate the player so that I can move the character with ease.
S2US8: As a player I want to be able to control the camera movement with the mouse so that I can play with ease.
S2US18: As a player I would prefer for the camera to be higher and further away so that I can see more of the map and do not have to rely on keys to move the camera.