S2US66: (fixing bugs)
S2US66: As a player, I want to be able to go through the main door in the level which will then let me go to the next level. Each level should be a rollover with the inventory scene between them so that I do not need to go to the level select if I do not want to. Estimated (2 hour)