S2US84:- complete and S2US23: merge bug error fixing
S2US84: As a player I want the painter models (orange, yellow and green) to be functional within the scene so that I can play the game and experience the theme. Estimated (1 hour)
and
S2US23: As a player I do not want there to be any errors when I run into the resources so that I can show collection. (40mins) - merge bug error