Implemented S3US321, S3US27, S3US28 + S3US24
S3US321: As a player I want the movement to be freer as I found it hard to control the movement whilst trying to switch colors and shoot enemies.
S3US27: As a player I want to aim to be separate from movement so that I can get through the level without experiencing difficulties.
S3US28: As a player I want a better control scheme to be used for hitting the enemies with the paint.
S3US24: As a player I want there to be able to aim where I want to shoot and this be the only use of the mouse so that I am not confused on the mouses ability.