S4US38:  and S4US39:  working on them
S4US38: As a designer I want to restructure the level layout of the four level, to contain all the new models and features so that the player is experiencing the correct level of challenge. 2 hours

and

S4US39: As a designer I want to restructure the level layout of the fifth level, to contain all the new models and features so that the player is experiencing the correct level of challenge. 2 hours
Jesica