S4US42: complete and S4US43: complete
S4US42: As a player, I want health items that I can either collect or step on which will give me health so that I have the option to save the character when in need. 30mins

S4US43: As a player I want the resources to be bigger and for them to disappear slower so that it flows with the game. 20mins